fixed missing eps in Animixplay
